app/requirements.txt
flask
flask_sqlalchemy
pymysql
python-dotenv
